# openapi.yaml
openapi: 3.0.3
info:
  title: RESIZR - Image Processing API
  description: |
    **RESIZR** is a high-performance image processing service that provides RESTful APIs for image upload, processing, and delivery. 
    
    ## Features
    - Multi-resolution image processing with configurable resize algorithms (smart-fit, crop, stretch)
    - Support for JPEG, PNG, GIF, and WebP formats
    - Redis-based metadata storage and caching
    - S3-compatible storage backend
    - Rate limiting and security features
    
    ## Authentication
    RESIZR supports optional API key-based authentication with two permission levels:
    
    - **Read-Write Keys**: Can upload images and access all read operations
    - **Read-Only Keys**: Can only access download and info operations
    
    Authentication is disabled by default and can be enabled via configuration.
    When enabled, API keys must be provided in the `X-API-Key` header (configurable).
    
    ## Rate Limits
    - Upload: 10 requests/minute per IP
    - Download: 100 requests/minute per IP
    - Info: 50 requests/minute per IP
    
    ## CORS Support
    Cross-Origin Resource Sharing (CORS) is configurable:
    - Production: Restricted to configured allowed origins
    - Development: Allows all origins for testing
    - Configurable through environment variables
    
    ## Error Handling
    All errors follow a consistent JSON format with appropriate HTTP status codes.
    
  version: 0.0.1
  contact:
...
  license:
...
servers:
...
paths:
  /api/v1/auth/generate-key:
    get:
      tags:
        - Authentication
      summary: Generate API key
      description: |
        Generate a new cryptographically secure API key.
        
        **Permission Assignment:** The key's permission level is determined by which environment variable it's added to:
        - Add to `AUTH_READWRITE_KEYS` for full access (uploads, downloads, info)
        - Add to `AUTH_READONLY_KEYS` for read-only access (downloads, info only)
        
        **Note:** Generated keys must be manually added to environment configuration to become active.
        
      operationId: generateAPIKey
      responses:
        '201':
          description: API key generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_key:
                    type: string
                    description: Generated API key (64-character hex string)
                    example: a1b2c3d4e5f6789012345678901234567890abcdef1234567890abcdef123456
                  message:
                    type: string
                    description: Success message with configuration instructions
                    example: API key generated successfully. Add it to AUTH_READWRITE_KEYS or AUTH_READONLY_KEYS environment variable to activate.
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key authentication. Required when authentication is enabled.
        
        Two types of keys are supported:
        - **Read-Write Keys**: Can upload images and access all read operations
        - **Read-Only Keys**: Can only access download and info operations
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - code
      properties:
        error:
          type: string
          description: Error category or type
          example: "Validation failed"
        message:
          type: string
          description: Human-readable error message
          example: "The uploaded file exceeds the maximum size limit"
        code:
          type: integer
          description: HTTP status code
          example: 400
...
security: []
...